rename LogEditor to ProcessLogEdit
